cleanup: rename garmin_sync to garmin_fetcher
Rename src/garmin_sync.py → src/garmin_fetcher.py and update all
import sites, script references, and tests. Rename only; no logic
changes to the module itself.

Updated references:
- bin/sync_garmin_data.sh: script invocation path
- bin/delete_duplicate_workouts.py: bare import via sys.path
- src/workout_uploader.py: dynamic import in get_garmin_client
- tests/test_garmin_sync.py: all from-imports (7 occurrences)
- skills/garmin_sync.py: docstring comment (accuracy)
